Extract Base Handler
