+++ .env.example
@@ -1,0 +1,3 @@
+OPENAI_TOKEN='from https://platform.openai.com/account/api-keys'
+TELEGRAM_BOT_TOKEN='ask @BotFather to create a bot and get the token'
+SKYNET_DATABASE_PASSWORD='production_db_password'
