+++ README.md
@@ -2,23 +2,20 @@ This is a simple Telegram bot, the idea of which I spied on in one of the chats.
 The bot can respond to text or voice messages.
 For this, Wisper speech recognition technologies and ChatGPT capabilities are used.
 
-### Version 1 TODO
-Done
+### Stage 3
+1. Migrate to a new server
+2. run Postgres in docker with dumps
+3. CD
 
-### Version 2 TODO
-1. Change Locale according to user's language
-10. Add gritting message
-2. Add DB (sqlite3)/file
-8. Add logger ( files)
-3. Refactoring
-
-### Version 3 TODO
-9. Add Error handling
-11. Add support for images
-3. Refactoring
+### Stage 4 
+1. Add Error handling for OpenAI client
+2. Add User model 
+3. Save AI statistics (tokens, requests, etc)
+4. Add User statistics (tokens, requests, etc)
 
 
 ### How to run
-1. `bundle exec puma` or `rerun puma` for development mode
-2. Visit https://web.telegram.org/k/#@KyleReese_bot and start chatting with the bot
+1. create `.env` file with from `.env.example`
+2. rails server`
+3. Visit your Bot and start chatting
