+++ config/application.rb
@@ -37,11 +37,7 @@ module SkyNet
     # in config/environments, which are processed later.
     #
     # config.time_zone = "Central Time (US & Canada)"
-    # config.eager_load_paths << Rails.root.join('app', 'services')
-    # config.autoload_paths << Rails.root.join('app', 'services')
-    # config.autoload_paths << "#{Rails.root}/app/services"
-    # config.eager_load_paths << "#{Rails.root}/app/services"
-    # config.autoload_paths += %W(#{config.root}/app/services)
+    # config.eager_load_paths << Rails.root.join("extras")
     # Only loads a smaller set of middleware suitable for API only apps.
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
